feat(context): create julie-context crate with ToolContext trait + relocate SpilloverStore/WorkspaceTarget (Phase 2 PR 2b T2b.1)
- Add crates/julie-context to workspace members and root Cargo.toml deps
- julie-context deps: julie-core + julie-index + async-trait + anyhow + blake3 (no pipeline, no runtime, no tools)
- src/spillover.rs: SpilloverStore/Format/Page relocated verbatim from src/tools/spillover/store.rs
- src/workspace_target.rs: WorkspaceTarget enum relocated from src/tools/navigation/resolution.rs:43
- src/tool_context.rs: #[async_trait] ToolContext with 18-method validated U5 surface
- src/lib.rs: flat re-exports for SpilloverStore, WorkspaceTarget, ToolContext
- tests/no_upward_deps.rs: source + manifest tripwire; allows core/index only; needle crate::workspace:: (with trailing ::) avoids false-positive on crate::workspace_target

Re-export shims (all ~9 SpilloverStore importers + ~21 WorkspaceTarget importers unchanged):
- src/tools/spillover/store.rs → pub use julie_context::{SpilloverFormat, SpilloverPage, SpilloverStore}
- src/tools/navigation/resolution.rs → removed local WorkspaceTarget def, added pub use julie_context::WorkspaceTarget

Gates: nextest -p julie-context --no-run PASS; nextest -p julie --no-run PASS; nextest -p julie-context (2/2 tripwire) PASS; cargo build PASS. Negative tests: julie_pipeline source plant → tripwire FAIL (correct); julie-pipeline manifest plant → manifest guard FAIL (correct).
