+++ .memories/2026-06-04/112644_3181.md
@@ -1,0 +1,78 @@
+---
+id: checkpoint_3181a852
+timestamp: 2026-06-04T11:26:44.484Z
+tags:
+  - feature
+  - refactor
+  - crate-split
+  - julie-rescue
+  - phase2b
+  - tool-context
+  - facade-trait
+  - spillover
+  - workspace-target
+  - dependency-direction
+git:
+  branch: julie-rescue-phase2b
+  commit: 30482c47
+  files:
+    - .codex/config.toml
+    - .miller/.symbols.db.init.lock
+    - .miller/indexer.lock
+    - .miller/logs/miller-20260531.jsonl
+    - .miller/logs/miller-20260603.jsonl
+    - .miller/logs/miller-35698-20260531.jsonl
+    - .miller/logs/miller-38072-20260531.jsonl
+    - .miller/logs/miller-43360-20260531.jsonl
+    - .miller/symbols.db
+    - .miller/symbols.db.julie-extract.lock
+    - Cargo.lock
+    - Cargo.toml
+    - crates/julie-context/Cargo.toml
+    - crates/julie-context/src/lib.rs
+    - crates/julie-context/src/spillover.rs
+    - crates/julie-context/src/tool_context.rs
+    - crates/julie-context/src/workspace_target.rs
+    - crates/julie-context/tests/no_upward_deps.rs
+    - src/tools/navigation/resolution.rs
+    - src/tools/spillover/store.rs
+summary: T2b.1 complete — crates/julie-context created
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+---
+
+## T2b.1 complete — crates/julie-context created
+
+Phase 2 PR 2b first step: the `julie-context` crate is fully built and all gates pass.
+
+### What was created
+- `crates/julie-context/Cargo.toml` — deps: julie-core + julie-index + async-trait + anyhow + blake3 (no pipeline, no runtime, no tools)
+- `crates/julie-context/src/lib.rs` — flat re-exports of SpilloverStore/Format/Page, ToolContext, WorkspaceTarget
+- `crates/julie-context/src/spillover.rs` — SpilloverStore relocated verbatim from `src/tools/spillover/store.rs`
+- `crates/julie-context/src/workspace_target.rs` — WorkspaceTarget enum relocated from `src/tools/navigation/resolution.rs:43`
+- `crates/julie-context/src/tool_context.rs` — ToolContext trait with 18 methods (U5 validated surface)
+- `crates/julie-context/tests/no_upward_deps.rs` — tripwire (source + manifest guards)
+
+### Re-export shims (importers unchanged)
+- `src/tools/spillover/store.rs` → `pub use julie_context::{SpilloverFormat, SpilloverPage, SpilloverStore}`
+- `src/tools/navigation/resolution.rs` → removed WorkspaceTarget def, added `pub use julie_context::WorkspaceTarget`
+
+### Tripwire needle fix
+- `crate::workspace` changed to `crate::workspace::` (trailing `::`) to avoid false-positive match on the legitimate in-crate `crate::workspace_target` module
+
+### Verification gates (all PASS)
+1. `cargo nextest run -p julie-context --no-run` — context crate incl. cfg(test) compiles
+2. `cargo nextest run -p julie --no-run` — top-crate test binary compiles after shims
+3. `cargo nextest run -p julie-context` — 2/2 tripwire tests pass
+4. `cargo build` — debug binary links (Finished dev profile, no warnings)
+
+### Negative tests (both fire)
+- Planted `julie_pipeline` in src file → `no_upward_source_references` FAILED with expected message
+- Planted `julie-pipeline = ...` in Cargo.toml → `manifest_has_no_cyclic_or_upward_dependency` FAILED with expected message
+- Both plants removed; final clean run: 2/2 green
+
+### Trait surface (18 methods, U5 validated)
+**Sync:** current_workspace_id, require_primary_workspace_identity, require_primary_workspace_root, loaded_workspace_id, is_primary_workspace_swap_in_progress, session_id() -> &str (NEW), spillover_store() -> Arc\<SpilloverStore\> (NEW)
+**Async:** primary_pooled_database, primary_pooled_database_and_search_index, get_pooled_database_for_workspace, get_database_for_workspace (NEW), get_search_index_for_workspace, get_workspace_root_for_target, embedding_provider, ensure_embedding_provider(timeout), resolve_workspace_target, ensure_target_workspace_indexed_if_pending, system_readiness
+
+### Next step
+T2b.2 — `impl ToolContext for JulieServerHandler` in the top crate (4 purpose-method impls + 14 accessor bodies verbatim)
