+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [".", "xtask", "crates/julie-core", "crates/julie-index", "crates/julie-pipeline", "crates/julie-test-support"]
+members = [".", "xtask", "crates/julie-core", "crates/julie-index", "crates/julie-pipeline", "crates/julie-context", "crates/julie-test-support"]
 resolver = "2"
 
 [workspace.package]
@@ -51,6 +51,9 @@ julie-index = { path = "crates/julie-index" }
 
 # Julie pipeline crate (indexing engine + embeddings, above julie-index)
 julie-pipeline = { path = "crates/julie-pipeline", features = ["embeddings-sidecar"] }
+
+# Julie context crate (ToolContext facade trait + SpilloverStore + WorkspaceTarget)
+julie-context = { path = "crates/julie-context" }
 
 # MCP Server SDK - Official Anthropic SDK
 rmcp = { version = "1.6", features = ["server", "client", "transport-io", "transport-streamable-http-server", "transport-streamable-http-client", "transport-streamable-http-client-reqwest", "macros"] }
